Get URL content based on if its Shiny or local

This changes the auth type depending on if its local or on Shiny

doHttrRequest(url, shiny_access_token = NULL, request_type = "GET",
  the_body = NULL, customConfig = NULL,
  simplifyVector = getOption("googleAuthR.jsonlite.simplifyVector"))

Arguments

url
the url of the page to retrieve

request_type
the type of httr request function: GET, POST, PUT, DELETE etc.

the_body
body of POST request

customConfig
list of httr options such as httr::use_proxy or httr::add_headers that will be added to the request.

simplifyVector
Passed to jsonlite::fromJSON

session
a shiny session object if in a shiny app.

params
A named character vector of other parameters to add to request.

Details

Example of params: c(param1="foo", param2="bar")