Authorize googleAuthR

Wrapper of token_fetch

gar_auth(
  token = NULL,
  email = NULL,
  scopes = getOption("googleAuthR.scopes.selected"),
  app = gar_oauth_app(),
  cache = gargle::gargle_oauth_cache(),
  use_oob = gargle::gargle_oob_default(),
  package = "googleAuthR",
  new_user = NULL
)

Arguments

token
an actual token object or the path to a valid token stored as an .rds file

email
An existing gargle cached email to authenticate with or TRUE to authenticate with the only email available.

scopes
Scope of the request

app
app as specified by gar_auth_configure

cache
Where to store authentication tokens

use_oob
Whther to use OOB browserless authetication

package
The name of the package authenticating

new_user
Deprecated, not used

Value

an OAuth token object, specifically a Token2.0, invisibly

See also

Examples

if (FALSE) {

# sets GCP project to auth through
gar_auth_configure(path="path/to/gcp-client.json")

# starts auth process with defaults
gar_auth()

# switching between auth scopes
# first time new scope manual auth, then auto if supplied email   
gar_auth(email = "your@email.com",
         scopes = "https://www.googleapis.com/auth/drive")

# ... query Google Drive functions ...

gar_auth(email = "your@email.com",
         scopes = "https://www.googleapis.com/auth/bigquery")

# ..query BigQuery functions ...

}