Perform auto authentication

This helper function lets you use environment variables to auto-authenticate on package load, intended for calling by gar_attach_auto_auth

gar_auto_auth(
  required_scopes,
  no_auto = NULL,
  environment_var = "GAR_AUTH_FILE",
  new_user = NULL
)

Arguments

required_scopes
Required scopes needed to authenticate - needs to match at least one

no_auto
If TRUE, ignore auto-authentication settings

environment_var
Name of environment var that contains auth file path

new_user
Deprecated, not used

The authentication file can be a .httr-oauth file created via gar_auth or a Google service JSON file downloaded from the Google API credential console, with file extension .json.

You can use this in your code to authenticate from a file location specified in file, but it is mainly intended to be called on package load via gar_attach_auto_auth.

environment_var This is the name that will be called via Sys.getenv on library load. The environment variable will contain an absolute file path to the location of an authentication file.

Value

an OAuth token object, specifically a Token2.0, invisibly

See also