Turn a list of gar_fetch_functions into batch functions

Turn a list of gar_fetch_functions into batch functions

gar_batch(call_list, ...)

Arguments

call_list

a list of functions from gar_api_generator

...

further arguments passed to the data parse function of f

Value

A list of the Google API responses

Details

This function will turn all the individual Google API functions into one POST request to /batch.

If you need to pass multiple data parse function arguments its probably best to do it in separate batches to avoid confusion.

See also

https://developers.google.com/webmaster-tools/v3/how-tos/batch

Documentation on doing batch requests for the search console API. Other Google APIs are similar.

Walk through API calls changing parameters using gar_batch_walk

Other batch functions: gar_batch_walk