gar_check_existing_token {googleAuthR}	R Documentation

Check a token vs options

Description

Useful for debugging authentication issues

Usage

gar_check_existing_token(token = .auth$cred)

Arguments

token
A token to check, default current live session token

Details

Will compare the passed token's settings and compare to set options. If these differ, then reauthentication may be needed.

Value

FALSE if the options and current token do not match, TRUE if they do.

[Package googleAuthR version 1.4.1 Index]