Create an API library skeleton

This will create a file with the skeleton of the API functions for the specified library

gar_create_api_skeleton(filename, api_json, format = TRUE)

Arguments

filename
R file to write skeleton to

api_json
The json from gar_discovery_api

format
If TRUE will use tidy_eval on content

Value

TRUE if successful, side effect will write a file

See also