Create a Google API package

Create a Google API package

gar_create_package(api_json, directory, rstudio = TRUE, check = TRUE,
  github = TRUE, format = TRUE, overwrite = TRUE)

Arguments

api_json
json from gar_discovery_api

directory
Where to build the package

rstudio
Passed to create, creates RStudio project file

check
Perform a check on the package once done

github
If TRUE will upload package to your github

format
If TRUE will use tidy_eval on content

overwrite
Whether to overwrite an existing directory if it exists

Value

If check is TRUE, the results of the CRAN check, else FALSE

Details

For github upload to work you need to have your github PAT setup. See use_github.

Uses devtools' create to create a package structure then gar_create_api_skeleton and gar_create_api_objects to create starting files for a Google API package.

See also

https://developers.google.com/discovery/v1/reference/apis/list

A Github repository with https://github.com/MarkEdmondson1234/autoGoogleAPI generated by this function.

Other Google Discovery API functions: gar_create_api_objects, gar_create_api_skeleton, gar_discovery_apis_list, gar_discovery_api