gar_debug_parsing {googleAuthR}	R Documentation

Read the diagnostic object returned on API parse errors.

Description

Read the diagnostic object returned on API parse errors.

Usage

gar_debug_parsing(filename = "gar_parse_error.rds")

Arguments

filename
The file created from API errors, usually called gar_parse_error.rds

Details

When googleAuthR API parsing fails, it will write a file called gar_parse_error.rds to the directory. Feed that file into this function to help diagnose the problem.

[Package googleAuthR version 1.4.1 Index]