gar_setup_auth_key {googleAuthR}	R Documentation

Create a service account for googleCloudRunner

Description

This will use your Google OAuth2 user to create a suitable service account

Usage

gar_setup_auth_key(
  email = Sys.getenv("GARGLE_EMAIL"),
  file = "googleauthr-auth-key.json",
  session_user = NULL,
  client_json = "GAR_CLIENT_JSON",
  roles = NULL,
  default_key = "googleauthr"
)

Arguments

email
What email to open OAuth2 with

file
Where to save the authentication file

session_user
1 for user level, 2 for project level, leave NULL to be prompted

client_json
The location of the env arg holding client json

roles
Whether to assign roles to the service key

default_key
The default name of the service key

Value

TRUE if the file is ready to be setup, FALSE if need to stop

See Also

Other setup functions: gar_setup_auth_check(), gar_setup_clientid(), gar_setup_edit_renviron(), gar_setup_env_check(), gar_setup_menu_do(), gar_setup_menu()

[Package googleAuthR version 1.4.1 Index]