gar_setup_edit_renviron {googleAuthR}	R Documentation

Setup wizard help - asking users to edit .Renviron

Description

Setup wizard help - asking users to edit .Renviron

Usage

gar_setup_edit_renviron(to_paste, session_user)

gar_setup_check_session(session_user = NULL)

Arguments

to_paste
The line to paste into .Renviron

session_user
whether its a 1 = user level or 2=project level .Renviron file

Intended to get user input from a menu, 1 indicating user level, 2 project level

gar_setup_check_session creates a menu for the user to choose which

See Also

Other setup functions: gar_setup_auth_check(), gar_setup_auth_key(), gar_setup_clientid(), gar_setup_env_check(), gar_setup_menu_do(), gar_setup_menu()

Examples

## Not run: 

choice <- gar_setup_check_session()
gar_setup_edit_renviron("ENV_ARG=blah", session_user = choice)

## End(Not run)

[Package googleAuthR version 1.4.1 Index]