Make a Google Authorisation URL for Shiny

Set this within your login_ui where you need the Google login.

gar_shiny_auth_url(req, state = getOption("googleAuthR.securitycode"),
  client.id = getOption("googleAuthR.webapp.client_id"),
  client.secret = getOption("googleAuthR.webapp.client_secret"),
  scope = getOption("googleAuthR.scopes.selected"),
  access_type = c("online", "offline"), approval_prompt = c("auto",
  "force"))

Arguments

req
a Rook request, do not set as this will be used by Shiny to generate URL

state
URL state

client.id
client.id

client.secret
client.secret

scope
API scopes

access_type
whether to keep the token

approval_prompt
Auto-login if user is recognised or always force signin

See also