Returns the authentication Token.

Once a user browses to ShinyGetTokenURL and is redirected back with request gar_shiny_getToken takes that code and returns a token needed for Google APIs Uses the same client.id and client.secret as ShinyGetTokenURL.

gar_shiny_getToken(code, redirect.uri,
  client.id = getOption("googleAuthR.webapp.client_id"),
  client.secret = getOption("googleAuthR.webapp.client_secret"))

Arguments

code

The code returned from a successful Google authentication.

redirect.uri

Where a user will go after authentication, from gar_shiny_getUrl(session)

client.id

From the Google API console.

client.secret

From the Google API console.

Value

A list including the token needed for Google API requests.

See also

Other shiny auth functions: authReturnCode, createCode, gar_shiny_getAuthUrl, gar_shiny_getUrl, loginOutput, reactiveAccessToken, renderLogin, revokeEventObserver, with_shiny