gar_shiny_login_ui {googleAuthR}	R Documentation

A login page for Shiny

Description

An alternative to the immediate login provided by default by gar_shiny_ui

Usage

gar_shiny_login_ui(req, title = "googleAuthR Login Demo")

Arguments

req
Passed to gar_shiny_auth_url to generate login URL

title
The title of the page

Details

Use gar_shiny_auth_url to create the login URL. You must leave the first argument free as this is used to generate the login, but you can pass other arguments to customise your UI.

See Also

Other pre-load shiny authentication: gar_shiny_auth_url(), gar_shiny_auth(), gar_shiny_ui(), silent_auth()

[Package googleAuthR version 1.4.1 Index]