gar_token {googleAuthR}	R Documentation

Produce configured token

Description

For internal use or for those programming around the Google API. Returns a token pre-processed with [httr::config()]. Most users do not need to handle tokens "by hand" or, even if they need some control, [gar_auth()] is what they need. If there is no current token, [gar_auth()] is called to either load from cache or initiate OAuth2.0 flow. If auth has been deactivated via [gar_deauth()], 'gar_token()' returns 'NULL'.

Usage

gar_token()

Value

A 'request' object (an S3 class provided by [httr][httr::httr]).

See Also

Other low-level API functions: gar_has_token()

Examples

## Not run: 
req <- request_generate(
  "drive.files.get",
  list(fileId = "abc"),
  token = gar_token()
)
req

## End(Not run)

[Package googleAuthR version 1.4.1 Index]