get_google_token {googleAuthR}	R Documentation

Retrieve Google token from environment and configs for httr

Description

Get token if it's previously stored, else prompt user to get one.

Usage

get_google_token(shiny_return_token = NULL)

Arguments

shiny_return_token
In a shiny session, this is passed instead.

Value

a httr configured option for token For shiny the token is passed from reactive session

See Also

Other authentication functions: gar_attach_auto_auth(), gar_auth_service(), gar_auth(), gar_auto_auth(), gar_gce_auth(), token_exists()

[Package googleAuthR version 1.4.1 Index]