silent_auth {googleAuthR}	R Documentation

Silent auth

Description

The default for logging in via gar_shiny_ui, this creates no login page and just takes you straight to authentication on Shiny app load.

Usage

silent_auth(req)

Arguments

req
What Shiny uses to check the URL parameters

See Also

Other pre-load shiny authentication: gar_shiny_auth_url(), gar_shiny_auth(), gar_shiny_login_ui(), gar_shiny_ui()

[Package googleAuthR version 1.4.1 Index]