skip_if_no_env_auth {googleAuthR}	R Documentation

Skip test if not authenticated

Description

Use within tests to skip if a local authentication file isn't available through an environment variable.

Usage

skip_if_no_env_auth(env_arg)

Arguments

env_arg
The name of the environment argument pointing to the auth file

[Package googleAuthR version 1.4.1 Index]