Auto Authentication function for use within .onAttach

To be placed within .onAttach to auto load an authentication file from an environment variable.

gar_attach_auto_auth(required_scopes, environment_var = "GAR_AUTH_FILE",
  travis_environment_var = NULL)

Arguments

required_scopes
A character vector of minimum required scopes for this API library

environment_var
The name of the environment variable where the filepath to the authentication file is kept

travis_environment_var
Defunct; now does nothing

This function works with gar_auto_auth. It is intended to be placed within the .onAttach hook so that it loads when you load your library.

For auto-authentication to work, the environment variable needs to hold a file path to an existing auth file such as created via gar_auth or a JSON file file download from the Google API console.

Value

Invisible, used for its side effects of calling auto-authentication.

See also

Other authentication functions: gar_auth_service, gar_auth, gar_auto_auth, gar_gce_auth, get_google_token, token_exists

Examples

# NOT RUN {
.onAttach <- function(libname, pkgname){

  googleAuthR::gar_attach_auto_auth("https://www.googleapis.com/auth/urlshortener", "US_AUTH_FILE")

}

## will only work if you have US_AUTH_FILE environment variable pointing to an auth file location
## .Renviron example
US_AUTH_FILE="/home/mark/auth/urlshortnerauth.json"

# }