gar_auto_auth {googleAuthR}	R Documentation

Perform auto authentication

Description

This helper function lets you use environment variables to auto-authenticate on package load, intended for calling by gar_attach_auto_auth

Usage

gar_auto_auth(required_scopes, environment_var = "GAR_AUTH_FILE")

Arguments

required_scopes
Required scopes needed to authenticate - needs to match at least one

environment_var
Name of environment var that contains auth file path

The authentication file can be a .httr-oauth file created via gar_auth or a Google service JSON file downloaded from the Google API credential console, with file extension .json.

You can use this in your code to authenticate from a file location specified in file, but it is mainly intended to be called on package load via gar_attach_auto_auth.

environment_var This is the name that will be called via Sys.getenv on library load. The environment variable will contain an absolute file path to the location of an authentication file.

Value

an OAuth token object, specifically a Token2.0, invisibly

See Also

Help files for .onAttach

Other authentication functions: gar_attach_auto_auth(), gar_auth_service(), gar_auth(), gar_gce_auth(), get_google_token(), token_exists()

[Package googleAuthR version 1.4.1 Index]