gar_create_api_skeleton {googleAuthR}	R Documentation

Create an API library skeleton

Description

This will create a file with the skeleton of the API functions for the specified library

Usage

gar_create_api_skeleton(filename, api_json, format = TRUE)

Arguments

filename
R file to write skeleton to

api_json
The json from gar_discovery_api

format
If TRUE will use tidy_eval on content

Value

TRUE if successful, side effect will write a file

See Also

Other Google Discovery API functions: gar_create_api_objects(), gar_create_package(), gar_discovery_apis_list(), gar_discovery_api()

[Package googleAuthR version 1.4.1 Index]