gar_create_package {googleAuthR}	R Documentation

Create a Google API package

Description

Create a Google API package

Usage

gar_create_package(
  api_json,
  directory,
  rstudio = TRUE,
  check = FALSE,
  github = FALSE,
  format = TRUE,
  overwrite = TRUE
)

Arguments

api_json
json from gar_discovery_api

directory
Where to build the package

rstudio
Passed to create_package, creates RStudio project file

check
Perform a check on the package once done

github
If TRUE will upload package to your github

format
If TRUE will use tidy_eval on content

overwrite
Whether to overwrite an existing directory if it exists

Details

For github upload to work you need to have your github PAT setup. See use_github.

Uses usethis to create a package structure then gar_create_api_skeleton and gar_create_api_objects to create starting files for a Google API package.

Value

If check is TRUE, the results of the CRAN check, else FALSE

See Also

https://developers.google.com/discovery/v1/reference/apis/list

A Github repository with 154 R packages examples generated by this function.

Other Google Discovery API functions: gar_create_api_objects(), gar_create_api_skeleton(), gar_discovery_apis_list(), gar_discovery_api()

[Package googleAuthR version 1.4.1 Index]