gar_gce_auth {googleAuthR}	R Documentation

Authenticate on Google Compute Engine

Description

This takes the metadata auth token in a Google Compute Engine instance as authentication source

Usage

gar_gce_auth(
  service_account = "default",
  scopes = "https://www.googleapis.com/auth/cloud-platform"
)

Arguments

service_account
Specify a different service account from the default

scopes
Scopes for the authentication

Details

service_account is default or the service account email e.g. "service-account-key-json@projectname.iam.gserviceaccount.com"

Google Compute Engine instances come with their own authentication tokens.

It has no refresh token so you need to call for a fresh token after approx. one hour. The metadata token will refresh itself when it has about 60 seconds left.

You can only use for scopes specified when creating the instance.

If you want to use them make sure their service account email is added to accounts you want to get data from.

Value

A token

See Also

gar_gce_auth_email

Other authentication functions: gar_attach_auto_auth(), gar_auth_service(), gar_auth(), gar_auto_auth(), get_google_token(), token_exists()

[Package googleAuthR version 1.4.1 Index]