gar_shiny_auth_url {googleAuthR}	R Documentation

Make a Google Authorisation URL for Shiny

Description

Set this within your login_ui where you need the Google login.

Usage

gar_shiny_auth_url(
  req,
  state = getOption("googleAuthR.securitycode"),
  client.id = getOption("googleAuthR.webapp.client_id"),
  client.secret = getOption("googleAuthR.webapp.client_secret"),
  scope = getOption("googleAuthR.scopes.selected"),
  access_type = c("online", "offline"),
  prompt = c("consent", "select_account", "both", "none")
)

Arguments

req
a Rook request, do not set as this will be used by Shiny to generate URL

state
URL state

client.id
client.id

client.secret
client.secret

scope
API scopes

access_type
whether to keep the token

prompt
Auto-login if user is recognised or always force signin

See Also

Other pre-load shiny authentication: gar_shiny_auth(), gar_shiny_login_ui(), gar_shiny_ui(), silent_auth()

[Package googleAuthR version 1.4.1 Index]